feat(video-editor): redesign timeline scrubber and unify add-layer flow
Rebuild the bottom timeline to match the design and dock it at the bottom
of the editor stage.

- Extract the timeline out of VideoJSPlayer into
  components/timeline/Timeline.js and share the layer-type registry via
  utils/layerTypes.js
- New track styling: accent progress fill, gray remainder, playhead
  handle, and a time ruler with major labels + minor tick dots
- Layer markers are always visible as icon chips and expand on hover to
  show "icon + type · time"; a selected layer stays collapsed
- Add a two-step "Add layer" affordance on the track: a "Click to add
  layer" hint follows the cursor, and clicking seeks + pins an "Add layer"
  chip
- Use a cart icon for the WooCommerce layer marker
- Remove the central LayerSelector modal; the timeline "Add layer" now
  opens the sidebar add-layer dropdown and inserts the layer at the
  clicked time
