+++ pages/video-editor/_controls.scss
@@ -35,6 +35,10 @@ $ve-radius: 4px;
 	.godam-ve-settings__head {
 		padding: 1rem;
 		border-bottom: 1px solid $ve-border;
+		position: sticky;
+		top: 0;
+		background: #fff;
+		z-index: 10;
 	}
 
 	.godam-ve-settings__title {
@@ -569,7 +573,10 @@ $ve-radius: 4px;
 		padding-bottom: 0.875rem;
 		margin-bottom: 1.25rem;
 		border-bottom: 1px solid $ve-border;
-
+		position: sticky;
+		top: 0;
+		background: #fff;
+		z-index: 10;
 
 		&__body {
 			display: flex;
