+++ pages/video-editor/_layers.scss
@@ -22,6 +22,10 @@ $ve-circle: color-mix(in srgb, var(--wp-components-color-accent, var(--wp-admin-
 		&__head {
 			padding: 1rem;
 			border-bottom: 1px solid $ve-border;
+			position: sticky;
+			top: 0;
+			background: #fff;
+			z-index: 10;
 		}
 
 		&__title {
